release/5.26.1

+++ core/datasource/src/main/java/com/tangem/datasource/api/stakekit/models/response/model/AddressArgument.kt
@@ -10,7 +10,7 @@ data class AddressArgument(
     @Json(name = "network")
     val network: String? = null,
     @Json(name = "minimum")
-    val minimum: Int? = null,
+    val minimum: Double? = null,
     @Json(name = "maximum")
-    val maximum: Int? = null,
+    val maximum: Double? = null,
 )
